test(button): Refactor visual testing stories to use CSF
